Rename test helper files to helpers_test.go
This ensures that these files are properly included only in testing.

[Fixes #10]
