+++ prometheus/helpers_test.go
@@ -38,3 +38,18 @@ func isNaN(obtained interface{}) (result bool) {
 
 	return false
 }
+
+type valueEqualsChecker struct {
+	*CheckerInfo
+}
+
+var ValueEquals Checker = &valueEqualsChecker{
+	&CheckerInfo{Name: "IsValue", Params: []string{"obtained", "expected"}},
+}
+
+func (checker *valueEqualsChecker) Check(params []interface{}, names []string) (result bool, error string) {
+	actual := params[0].(*item).Value
+	expected := params[1]
+
+	return actual == expected, ""
+}
